Fix loop variable ref issue in ufunc benchmarks.
Fix an issue where the loop variable (the ufunc name) was referenced
from within the constructed benchmark classes, which meant all of the
classes ended up running the same ufunc -- the last one in the list --
rather than whichever one they were supposed to be benchmarking.
The ufunc is now assigned to a class attribute on each of the benchmark
classes, and each class retrieves it from there avoiding capture issues.

Defined __slots__ for the constructed ufunc benchmark classes to shave
off a little overhead (from attribute lookups) within the benchmark
methods. It's likely negligible, but every little bit of overhead that
can be removed helps make the benchmark results as accurate as possible.